+++ app/client/src/ee/components/DSDataFilter/index.tsx
@@ -1,3 +1,3 @@
-import DSDataFilter from "ce/components/DSDataFilter";
-
-export default DSDataFilter;
+export * from "ce/components/DSDataFilter";
+import { default as CE_DSDataFilter } from "ce/components/DSDataFilter";
+export default CE_DSDataFilter;
