+++ app/client/src/ee/components/EnvConfigSection/index.tsx
@@ -1,18 +1,1 @@
-import type { Datasource } from "entities/Datasource";
-import { renderDatasourceSection } from "pages/Editor/DataSourceEditor/DatasourceSection";
-
-interface Props {
-  currentEnv: string;
-  config: any;
-  datasource: Datasource;
-  viewMode: boolean | undefined;
-}
-
-export function EnvConfigSection({
-  config,
-  currentEnv,
-  datasource,
-  viewMode,
-}: Props) {
-  return renderDatasourceSection(config, currentEnv, datasource, viewMode);
-}
+export * from "ce/components/EnvConfigSection";
