+++ app/client/src/ee/components/gitComponents/DisableCDModal.tsx
@@ -1,2 +1,3 @@
-import DisableCDModal from "ce/components/gitComponents/DisableCDModal";
-export default DisableCDModal;
+export * from "ce/components/gitComponents/DisableCDModal";
+import { default as CE_DisableCDModal } from "ce/components/gitComponents/DisableCDModal";
+export default CE_DisableCDModal;
