+++ app/client/src/ee/components/gitComponents/GitSettingsCDTab/index.tsx
@@ -1,2 +1,3 @@
-import GitSettingsCDTab from "ce/components/gitComponents/GitSettingsCDTab";
-export default GitSettingsCDTab;
+export * from "ce/components/gitComponents/GitSettingsCDTab";
+import { default as CE_GitSettingsCDTab } from "ce/components/gitComponents/GitSettingsCDTab";
+export default CE_GitSettingsCDTab;
