+++ app/client/src/ee/components/gitComponents/ReconfigureCDKeyModal.tsx
@@ -1,2 +1,3 @@
-import ReconfigureCDKeyModal from "ce/components/gitComponents/ReconfigureCDKeyModal";
-export default ReconfigureCDKeyModal;
+export * from "ce/components/gitComponents/ReconfigureCDKeyModal";
+import { default as CE_ReconfigureCDKeyModal } from "ce/components/gitComponents/ReconfigureCDKeyModal";
+export default CE_ReconfigureCDKeyModal;
