+++ app/client/src/ee/pages/AdminSettings/config/index.ts
@@ -1,3 +1,3 @@
 export * from "ce/pages/AdminSettings/config";
-import { ConfigFactory } from "pages/AdminSettings/config/ConfigFactory";
-export default ConfigFactory;
+import { default as CE_ConfigFactory } from "pages/AdminSettings/config/ConfigFactory";
+export default CE_ConfigFactory;
