+++ app/client/src/ee/plugins/Linting/lib/entity/index.ts
@@ -1,3 +1,3 @@
 export * from "ce/plugins/Linting/lib/entity/index";
-import EntityFactory from "ce/plugins/Linting/lib/entity";
-export default EntityFactory;
+import { default as CE_EntityFactory } from "ce/plugins/Linting/lib/entity";
+export default CE_EntityFactory;
