+++ app/client/src/ee/plugins/Linting/utils/getEntityUniqueIdForLogs.ts
@@ -1,4 +1,3 @@
 export * from "ce/plugins/Linting/utils/getEntityUniqueIdForLogs";
-import getEntityUniqueIdForLogs from "ce/plugins/Linting/utils/getEntityUniqueIdForLogs";
-
-export default getEntityUniqueIdForLogs;
+import { default as CE_getEntityUniqueIdForLogs } from "ce/plugins/Linting/utils/getEntityUniqueIdForLogs";
+export default CE_getEntityUniqueIdForLogs;
