+++ app/client/src/ee/plugins/Linting/utils/isLintErrorLoggingEnabledForEntity.ts
@@ -1,4 +1,3 @@
 export * from "ce/plugins/Linting/utils/isLintErrorLoggingEnabledForEntity";
-import isLintErrorLoggingEnabledForEntity from "ce/plugins/Linting/utils/isLintErrorLoggingEnabledForEntity";
-
-export default isLintErrorLoggingEnabledForEntity;
+import { default as CE_isLintErrorLoggingEnabledForEntity } from "ce/plugins/Linting/utils/isLintErrorLoggingEnabledForEntity";
+export default CE_isLintErrorLoggingEnabledForEntity;
