+++ app/client/src/ee/reducers/entityReducers/actionsReducer.tsx
@@ -1,4 +1,3 @@
 export * from "ce/reducers/entityReducers/actionsReducer";
 import { default as CE_actionsReducer } from "ce/reducers/entityReducers/actionsReducer";
-
 export default CE_actionsReducer;
