+++ app/client/src/ee/reducers/entityReducers/jsActionsReducer.tsx
@@ -1,4 +1,3 @@
 export * from "ce/reducers/entityReducers/jsActionsReducer";
 import { default as CE_jsActionsReducer } from "ce/reducers/entityReducers/jsActionsReducer";
-
 export default CE_jsActionsReducer;
